fixed onboarding for split view
