+++ Core/Sources/Core/Notification+Names.swift
@@ -10,5 +10,7 @@ import Foundation
 public extension Notification.Name {
     
     static let updateFavorites = Notification.Name("Notification.updateFavorites")
+    static let updateDashboard = Notification.Name("updateDashboard")
+    static let SetupDidComplete = Notification.Name("SetupDidCompleteNotification")
     
 }
