+++ Moers/Common/MainSplitViewController.swift
@@ -9,26 +9,49 @@
 import Foundation
 import UIKit
 import AppScaffold
+import MMUI
 import MMAPI
 import MMEvents
 import OSLog
+import BLTNBoard
+import Resolver
+import RubbishFeature
+import Combine
+import CoreLocation
+import Gestalt
 
 /// The main view controller which holds
 public class MainSplitViewController: UISplitViewController, SidebarViewControllerDelegate, UISplitViewControllerDelegate {
     
+    @LazyInjected var rubbishService: RubbishService
+    
     private let logger: Logger = Logger(.default)
     
+    var petrolManager: PetrolManagerProtocol
+    let locationManager: LocationManagerProtocol
     let dashboard: DashboardCoordinator
     let news: NewsCoordinator
     let map: MapCoordintor
     let events: EventCoordinator
     let other: OtherCoordinator
     
+    var firstLaunch: FirstLaunch
+    
     internal let tabController: TabBarController
     internal let sidebarController: SidebarViewController
     internal let secondaryRootViewControllers: [UIViewController]
     
+    deinit {
+        NotificationCenter.default.removeObserver(self)
+    }
+    
+    private var cancellables = Set<AnyCancellable>()
+    
     public var onChangeTraitCollection: ((UITraitCollection) -> Void)?
+    
+    public var displayCompact: Bool {
+        return self.traitCollection.horizontalSizeClass == .compact
+    }
     
     public init(
         firstLaunch: FirstLaunch,
@@ -39,6 +62,10 @@ public class MainSplitViewController: UISplitViewController, SidebarViewControll
         parkingLotManager: ParkingLotManagerProtocol,
         eventService: EventServiceProtocol
     ) {
+        
+        self.firstLaunch = firstLaunch
+        self.locationManager = locationManager
+        self.petrolManager = petrolManager
         
         self.dashboard = DashboardCoordinator(
             petrolManager: petrolManager
@@ -90,10 +117,38 @@ public class MainSplitViewController: UISplitViewController, SidebarViewControll
         
         self.configureSplitController()
         
+        if isSnapshotting() {
+            self.setupMocked()
+        }
+        
     }
     
     public required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
+    }
+    
+    // MARK: - UIViewController Lifecycle -
+    
+    public override func viewWillAppear(_ animated: Bool) {
+        super.viewWillAppear(animated)
+        
+        self.setupTheming()
+        self.loadRubbishData()
+        
+    }
+    
+    public override func viewDidAppear(_ animated: Bool) {
+        super.viewDidAppear(animated)
+        
+        self.firstLaunch = FirstLaunch(
+            userDefaults: .appGroup,
+            key: Constants.firstLaunch
+        )
+        
+        if (firstLaunch.isFirstLaunch || !onboardingManager.userDidCompleteSetup) && !isSnapshotting() {
+            showBulletin()
+        }
+        
     }
     
     private func configureSplitController() {
@@ -118,6 +173,31 @@ public class MainSplitViewController: UISplitViewController, SidebarViewControll
         
     }
     
+    private func setupTheming() {
+        
+        MMUIConfig.themeManager?.manage(theme: \Theme.self, for: self)
+        
+    }
+    
+    private func isSnapshotting() -> Bool {
+        return UserDefaults.standard.bool(forKey: "FASTLANE_SNAPSHOT")
+    }
+    
+    private func setupMocked() {
+        
+        UserManager.shared.register(User(type: .citizen, id: nil, name: nil, description: nil))
+        petrolManager.petrolType = .diesel
+        
+    }
+    
+    // MARK: - UISplitViewControllerDelegate -
+    
+    public func splitViewController(_ svc: UISplitViewController, willChangeTo displayMode: UISplitViewController.DisplayMode) {
+        
+        self.logger.info("MainSplitViewController changes to \(displayMode.rawValue, privacy: .public)")
+        
+    }
+    
     // MARK: - Sidebar Handling -
     
     public func sidebar(_ sidebarViewController: SidebarViewController, didSelectTabItem item: SidebarItem) {
@@ -134,21 +214,7 @@ public class MainSplitViewController: UISplitViewController, SidebarViewControll
         
     }
     
-    public func splitViewController(_ svc: UISplitViewController, willChangeTo displayMode: UISplitViewController.DisplayMode) {
-        
-        self.logger.info("MainSplitViewController changes to \(displayMode.rawValue, privacy: .public)")
-        
-    }
-    
-    public func switchToToday() {
-        selectSidebarItem(.dashboard)
-    }
-    
-    public func switchToNews() {
-        selectSidebarItem(.news)
-    }
-    
-    internal func selectSidebarItem(_ item: SidebarItem) {
+    public func selectSidebarItem(_ item: SidebarItem) {
         
         self.preferredDisplayMode = .oneBesideSecondary
         self.presentsWithGesture = false
@@ -162,10 +228,6 @@ public class MainSplitViewController: UISplitViewController, SidebarViewControll
         
     }
     
-    public var displayCompact: Bool {
-        return self.traitCollection.horizontalSizeClass == .compact
-    }
-    
     public override func traitCollectionDidChange(_ previousTraitCollection: UITraitCollection?) {
         
         if let previousTraitCollection = previousTraitCollection {
@@ -176,6 +238,185 @@ public class MainSplitViewController: UISplitViewController, SidebarViewControll
             }
             
         }
+        
+    }
+    
+    // MARK: - Onboarding -
+    
+    lazy var onboardingManager: OnboardingManager = {
+        return OnboardingManager()
+    }()
+    
+    lazy var bulletinManager: BLTNItemManager = {
+        let onboarding = onboardingManager.makeOnboarding()
+        return BLTNItemManager(rootItem: onboarding)
+    }()
+    
+    lazy var rubbishMigrationManager: BLTNItemManager = {
+        let page = makeRubbishMigrationPage()
+        return BLTNItemManager(rootItem: page)
+    }()
+    
+    // MARK: - Actions -
+    
+    public func updateDashboard() {
+        if self.isCollapsed {
+            tabController.dashboard.updateUI()
+        } else {
+            dashboard.updateUI()
+        }
+    }
+    
+    @objc func setupDidComplete() {
+        
+        AnalyticsManager.shared.logCompletedOnboarding()
+        
+        self.onboardingManager.userDidCompleteSetup = true
+        self.loadCurrentLocation()
+        self.updateDashboard()
+        
+    }
+    
+    private func showBulletin() {
+        
+        NotificationCenter.default.addObserver(self, selector: #selector(setupDidComplete), name: .SetupDidComplete, object: nil)
+        
+        bulletinManager.backgroundViewStyle = .dimmed
+        bulletinManager.statusBarAppearance = .hidden
+        bulletinManager.showBulletin(above: self)
+        
+    }
+    
+    private func makeRubbishMigrationPage() -> BLTNPageItem {
+        
+        let page = onboardingManager.makeRubbishStreetPage()
+        page.descriptionText = "Wähle Deine Straße erneut aus, um die aktuellen Abfuhrtermine der Müllabfuhr angezeigt zu bekommen."
+        
+        page.actionHandler = { [weak self] item in
+            
+            guard let item = item as? RubbishStreetPickerItem else { return }
+            
+            if var rubbishService = self?.rubbishService {
+                
+                rubbishService.register(item.selectedStreet)
+                rubbishService.isEnabled = true
+                
+                if rubbishService.remindersEnabled {
+                    rubbishService.registerNotifications(
+                        at: rubbishService.reminderHour ?? 20,
+                        minute: rubbishService.reminderHour ?? 00
+                    )
+                }
+                
+            }
+            
+            item.manager?.dismissBulletin(animated: true)
+            
+            self?.updateDashboard()
+            
+        }
+        
+        page.alternativeHandler = { [weak self] item in
+            
+            self?.rubbishService.isEnabled = false
+            self?.rubbishService.remindersEnabled = false
+            self?.rubbishService.disableReminder()
+            
+            item.manager?.dismissBulletin(animated: true)
+            
+            self?.updateDashboard()
+            
+        }
+        
+        return page
+        
+    }
+    
+    // MARK: - Data Handling
+    
+    private func loadCurrentLocation() {
+        
+        locationManager.authorizationStatus.sink { (authorizationStatus: CLAuthorizationStatus) in
+            if authorizationStatus == .authorizedWhenInUse {
+                self.locationManager.requestCurrentLocation()
+            }
+        }
+        .store(in: &cancellables)
+        
+    }
+    
+    private func loadRubbishData() {
+        
+        let rubbishService: RubbishService? = Resolver.optional()
+        
+        guard let rubbishService = rubbishService else {
+            return
+        }
+        
+        if rubbishService.isEnabled &&
+            !firstLaunch.isFirstLaunch &&
+            onboardingManager.userDidCompleteSetup &&
+            rubbishService.rubbishStreet != nil {
+            
+            rubbishService.loadRubbishCollectionStreets()
+                .receive(on: DispatchQueue.main)
+                .sink { (_: Subscribers.Completion<Error>) in
+                    
+                } receiveValue: { (streets: [RubbishFeature.RubbishCollectionStreet]) in
+                    
+                    let currentStreetName = rubbishService.rubbishStreet?.street ?? ""
+                    let currentStreetAddition = rubbishService.rubbishStreet?.streetAddition
+                    
+                    if let filteredStreet = streets.filter({
+                        $0.street == currentStreetName &&
+                        $0.streetAddition == currentStreetAddition
+                    }).first {
+                        
+                        rubbishService.register(filteredStreet)
+                        
+                        if rubbishService.remindersEnabled {
+                            rubbishService.registerNotifications(
+                                at: rubbishService.reminderHour ?? 20,
+                                minute: rubbishService.reminderMinute ?? 00
+                            )
+                        }
+                        
+                    } else {
+                        
+                        rubbishService.disableStreet()
+                        
+                        self.rubbishMigrationManager.backgroundViewStyle = .dimmed
+                        self.rubbishMigrationManager.statusBarAppearance = .hidden
+                        self.rubbishMigrationManager.showBulletin(above: self)
+                        
+                    }
+                    
+                }
+                .store(in: &cancellables)
+            
+        }
+        
+    }
+    
+}
+
+extension MainSplitViewController: Themeable {
+    
+    public typealias Theme = ApplicationTheme
+    
+    public func apply(theme: Theme) {
+        
+        self.view.backgroundColor = theme.backgroundColor
+        self.bulletinManager.backgroundColor = theme.backgroundColor
+        self.bulletinManager.hidesHomeIndicator = false
+        self.bulletinManager.edgeSpacing = .compact
+        self.rubbishMigrationManager.backgroundColor = theme.backgroundColor
+        self.rubbishMigrationManager.hidesHomeIndicator = false
+        self.rubbishMigrationManager.edgeSpacing = .compact
+        
+        let barAppearance = UIBarAppearance()
+        barAppearance.configureWithDefaultBackground()
+        barAppearance.backgroundColor = UIColor.systemBackground
         
     }
