+++ Moers/Controller/TabBarController.swift
@@ -38,20 +38,6 @@ public class TabBarController: UITabBarController, UITabBarControllerDelegate {
     
     private var cancellables = Set<AnyCancellable>()
     
-    lazy var onboardingManager: OnboardingManager = {
-        return OnboardingManager()
-    }()
-    
-    lazy var bulletinManager: BLTNItemManager = {
-        let onboarding = onboardingManager.makeOnboarding()
-        return BLTNItemManager(rootItem: onboarding)
-    }()
-    
-    lazy var rubbishMigrationManager: BLTNItemManager = {
-        let page = makeRubbishMigrationPage()
-        return BLTNItemManager(rootItem: page)
-    }()
-    
     init(
         firstLaunch: FirstLaunch,
         locationManager: LocationManagerProtocol,
@@ -94,10 +80,6 @@ public class TabBarController: UITabBarController, UITabBarControllerDelegate {
         
         super.init(nibName: nil, bundle: nil)
         
-        if isSnapshotting() {
-            self.setupMocked()
-        }
-        
     }
     
     deinit {
@@ -126,23 +108,11 @@ public class TabBarController: UITabBarController, UITabBarControllerDelegate {
         self.tabBar.accessibilityIdentifier = AccessibilityIdentifiers.tabBar
         
     }
-
+    
     public override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
         
         self.setupTheming()
-        self.loadRubbishData()
-        
-    }
-    
-    public override func viewDidAppear(_ animated: Bool) {
-        super.viewDidAppear(animated)
-        
-        self.firstLaunch = FirstLaunch(userDefaults: .appGroup, key: Constants.firstLaunch)
-        
-        if (firstLaunch.isFirstLaunch || !onboardingManager.userDidCompleteSetup) && !isSnapshotting() {
-            showBulletin()
-        }
         
     }
     
@@ -151,81 +121,6 @@ public class TabBarController: UITabBarController, UITabBarControllerDelegate {
     private func setupTheming() {
         
         MMUIConfig.themeManager?.manage(theme: \Theme.self, for: self)
-        
-    }
-    
-    // MARK: - Actions
-    
-    public func updateDashboard() {
-        dashboard.updateUI()
-    }
-    
-    @objc func setupDidComplete() {
-        
-        AnalyticsManager.shared.logCompletedOnboarding()
-        
-        self.onboardingManager.userDidCompleteSetup = true
-        self.loadCurrentLocation()
-        self.updateDashboard()
-        
-    }
-    
-    @objc func search() {
-        map.showSearch()
-    }
-    
-    private func showBulletin() {
-        
-        NotificationCenter.default.addObserver(self, selector: #selector(setupDidComplete), name: .SetupDidComplete, object: nil)
-        
-        bulletinManager.backgroundViewStyle = .dimmed
-        bulletinManager.statusBarAppearance = .hidden
-        bulletinManager.showBulletin(above: self)
-        
-    }
-    
-    private func makeRubbishMigrationPage() -> BLTNPageItem {
-        
-        let page = onboardingManager.makeRubbishStreetPage()
-        page.descriptionText = "Wähle Deine Straße erneut aus, um die aktuellen Abfuhrtermine der Müllabfuhr angezeigt zu bekommen."
-        
-        page.actionHandler = { [weak self] item in
-            
-            guard let item = item as? RubbishStreetPickerItem else { return }
-            
-            if var rubbishService = self?.rubbishService {
-                
-                rubbishService.register(item.selectedStreet)
-                rubbishService.isEnabled = true
-                
-                if rubbishService.remindersEnabled {
-                    rubbishService.registerNotifications(
-                        at: rubbishService.reminderHour ?? 20,
-                        minute: rubbishService.reminderHour ?? 00
-                    )
-                }
-                
-            }
-            
-            item.manager?.dismissBulletin(animated: true)
-            
-            self?.updateDashboard()
-            
-        }
-        
-        page.alternativeHandler = { [weak self] item in
-            
-            self?.rubbishService.isEnabled = false
-            self?.rubbishService.remindersEnabled = false
-            self?.rubbishService.disableReminder()
-            
-            item.manager?.dismissBulletin(animated: true)
-            
-            self?.updateDashboard()
-            
-        }
-        
-        return page
         
     }
     
@@ -242,64 +137,7 @@ public class TabBarController: UITabBarController, UITabBarControllerDelegate {
         
     }
     
-    private func loadRubbishData() {
-        
-        let rubbishService: RubbishService? = Resolver.optional()
-        
-        guard let rubbishService = rubbishService else {
-            return
-        }
-        
-        if rubbishService.isEnabled &&
-            !firstLaunch.isFirstLaunch &&
-            onboardingManager.userDidCompleteSetup &&
-            rubbishService.rubbishStreet != nil {
-            
-            rubbishService.loadRubbishCollectionStreets()
-                .receive(on: DispatchQueue.main)
-                .sink { (_: Subscribers.Completion<Error>) in
-                    
-                } receiveValue: { (streets: [RubbishFeature.RubbishCollectionStreet]) in
-                    
-                    let currentStreetName = rubbishService.rubbishStreet?.street ?? ""
-                    let currentStreetAddition = rubbishService.rubbishStreet?.streetAddition
-                    
-                    if let filteredStreet = streets.filter({
-                        $0.street == currentStreetName &&
-                        $0.streetAddition == currentStreetAddition
-                    }).first {
-                        
-                        rubbishService.register(filteredStreet)
-                        
-                        if rubbishService.remindersEnabled {
-                            rubbishService.registerNotifications(
-                                at: rubbishService.reminderHour ?? 20,
-                                minute: rubbishService.reminderMinute ?? 00
-                            )
-                        }
-                        
-                    } else {
-                        
-                        rubbishService.disableStreet()
-                        
-                        self.rubbishMigrationManager.backgroundViewStyle = .dimmed
-                        self.rubbishMigrationManager.statusBarAppearance = .hidden
-                        self.rubbishMigrationManager.showBulletin(above: self)
-                        
-                    }
-                    
-                }
-                .store(in: &cancellables)
-            
-        }
-        
-    }
-    
     // MARK: - Helper
-    
-    private func isSnapshotting() -> Bool {
-        return UserDefaults.standard.bool(forKey: "FASTLANE_SNAPSHOT")
-    }
     
     private func setupMocked() {
         
@@ -312,16 +150,6 @@ public class TabBarController: UITabBarController, UITabBarControllerDelegate {
         return UIStatusBarStyle.lightContent
     }
     
-    // MARK: - State Restoration
-    
-    public override func encodeRestorableState(with coder: NSCoder) {
-        super.encodeRestorableState(with: coder)
-    }
-    
-    public override func decodeRestorableState(with coder: NSCoder) {
-        super.decodeRestorableState(with: coder)
-    }
-    
 }
 
 extension TabBarController: Themeable {
@@ -330,22 +158,15 @@ extension TabBarController: Themeable {
     
     public func apply(theme: Theme) {
         
-//        UIApplication.shared.statusBarStyle = theme.statusBarStyle
         self.view.backgroundColor = theme.backgroundColor
         self.tabBar.tintColor = theme.accentColor
-        self.tabBar.barTintColor = UIColor.systemBackground // UIColor.black //theme.navigationBarColor
-        self.bulletinManager.backgroundColor = theme.backgroundColor
-        self.bulletinManager.hidesHomeIndicator = false
-        self.bulletinManager.edgeSpacing = .compact
-        self.rubbishMigrationManager.backgroundColor = theme.backgroundColor
-        self.rubbishMigrationManager.hidesHomeIndicator = false
-        self.rubbishMigrationManager.edgeSpacing = .compact
+        self.tabBar.barTintColor = UIColor.systemBackground
         
         self.tabBar.barStyle = .black
         
         let barAppearance = UIBarAppearance()
         barAppearance.configureWithDefaultBackground()
-        barAppearance.backgroundColor = UIColor.systemBackground // UIColor.black
+        barAppearance.backgroundColor = UIColor.systemBackground
         
         self.tabBar.standardAppearance = UITabBarAppearance(barAppearance: barAppearance)
